refactor: add in some function files and remove unit count when no units
