+++ src/node-lib/curriculum-api-2023/queries/unitListing/unitListing.query.ts
@@ -1,32 +1,15 @@
-import {
-  syntheticUnitvariantLessonsSchema,
-  yearSlugs,
-  yearDescriptions,
-} from "@oaknational/oak-curriculum-schema";
-import { kebabCase } from "lodash";
-import { z } from "zod";
+import { syntheticUnitvariantLessonsSchema } from "@oaknational/oak-curriculum-schema";
 
 import OakError from "../../../../errors/OakError";
 import { Sdk } from "../../sdk";
 
 import { getTiersForProgramme } from "./tiers/getTiersForProgramme";
 import { getUnitsForProgramme } from "./units/getUnitsForProgramme";
-import { getAllLearningThemes } from "./threads/getAllLearningThemes";
+import { getAllLearningThemes } from "./filters/getAllLearningThemes";
+import { getAllCategories } from "./filters/getAllCategroies";
+import { getAllYearGroups } from "./filters/getAllYearGroups";
 
 import { NEW_COHORT } from "@/config/cohort";
-
-// category icons don't have a consistent naming convention
-const categoryIconMap = {
-  Grammar: "subject-english-grammar",
-  Handwriting: "subject-english-handwriting",
-  "Reading, writing & oracy": "subject-english-reading-writing-oracy",
-  Spelling: "subject-english-spelling",
-  Vocabulary: "subject-english-vocabulary",
-  Physics: "subject-physics",
-  Biology: "subject-biology",
-  Chemistry: "subject-chemistry",
-};
-export type CategoryKeys = keyof typeof categoryIconMap;
 
 const unitListingQuery =
   (sdk: Sdk) => async (args: { programmeSlug: string }) => {
@@ -67,46 +50,11 @@ const unitListingQuery =
 
     const units = await getUnitsForProgramme(parsedRawUnits);
 
-    const yearGroups = Array.from(
-      units
-        .reduce((acc, unit) => {
-          const yearTitle = yearDescriptions.parse(unit[0]?.yearTitle);
-          const yearSlug = yearSlugs.parse(unit[0]?.year);
-
-          if (yearTitle && yearSlug && !acc.has(yearTitle)) {
-            acc.set(yearTitle, {
-              yearTitle: yearTitle,
-              year: yearSlug,
-            });
-          }
-          return acc;
-        }, new Map<z.infer<typeof yearDescriptions>, { yearTitle: z.infer<typeof yearDescriptions>; year: z.infer<typeof yearSlugs> }>())
-        .values(),
-    );
+    const yearGroups = getAllYearGroups(units);
 
     const learningThemes = getAllLearningThemes(units);
 
-    const subjectCategories = Array.from(
-      parsedRawUnits
-        .reduce((acc, unit) => {
-          if (!unit.unit_data.subjectcategories) {
-            return acc;
-          }
-          (unit.unit_data.subjectcategories as CategoryKeys[]).forEach(
-            (category) => {
-              if (typeof category === "string" && !acc.has(category)) {
-                acc.set(category, {
-                  label: category,
-                  iconName: categoryIconMap[category],
-                  slug: kebabCase(category),
-                });
-              }
-            },
-          );
-          return acc;
-        }, new Map<string, { label: string; iconName: string; slug: string }>())
-        .values(),
-    ).sort((a, b) => a.label.localeCompare(b.label));
+    const subjectCategories = getAllCategories(parsedRawUnits);
 
     const hasNewContent = units
       .flatMap((unit) => unit.flatMap((u) => u.cohort ?? "2020-2023"))
